+++ Tests/SqlCommandBuilder.Tests.ps1
@@ -18,36 +18,36 @@ Describe "SqlCommandBuilder" {
 
 	Context "GetDeleteCommand" {
 		It "should return the SQL command to delete an entity" {
-			$command = [SqlCommandBuilder]::new($connection).GetDeleteCommand($record).Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetDeleteCommand($record)
 			$command.Text | Should -BeLikeExactly 'DELETE FROM "main"."Characters"*'
 			$command.Text | Should -BeLikeExactly '*WHERE "ID" = @ID'
 		}
 
 		It "should also return the parameters used by the SQL command" {
-			$parameter = [SqlCommandBuilder]::new($connection).GetDeleteCommand($record).Item2[0]
-			$parameter.Name | Should -BeExactly "@ID"
-			$parameter.Value | Should -Be 1000
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetDeleteCommand($record)
+			$parameters[0].Name | Should -BeExactly "@ID"
+			$parameters[0].Value | Should -Be 1000
 		}
 	}
 
 	Context "GetExistsCommand" {
 		It "should return the SQL command to check the existence of an entity" {
-			$command = [SqlCommandBuilder]::new($connection).GetExistsCommand([Character], $record.Id).Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetExistsCommand([Character], $record.Id)
 			$command.Text | Should -BeLikeExactly "SELECT 1*"
 			$command.Text | Should -BeLikeExactly '*FROM "main"."Characters"*'
 			$command.Text | Should -BeLikeExactly '*WHERE "ID" = @ID'
 		}
 
 		It "should also return the parameters used by the SQL command" {
-			$parameter = [SqlCommandBuilder]::new($connection).GetExistsCommand([Character], $record.Id).Item2[0]
-			$parameter.Name | Should -BeExactly "@ID"
-			$parameter.Value | Should -Be 1000
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetExistsCommand([Character], $record.Id)
+			$parameters[0].Name | Should -BeExactly "@ID"
+			$parameters[0].Value | Should -Be 1000
 		}
 	}
 
 	Context "GetFindCommand" {
 		It "should return the SQL command to find an entity" {
-			$command = [SqlCommandBuilder]::new($connection).GetFindCommand([Character], $record.Id).Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetFindCommand([Character], $record.Id)
 			$command.Text | Should -BeLikeExactly 'SELECT "*'
 			$command.Text | Should -Not -BeLike '*`**'
 			$command.Text | Should -BeLikeExactly '*FROM "main"."Characters"*'
@@ -55,13 +55,13 @@ Describe "SqlCommandBuilder" {
 		}
 
 		It "should also return the parameters used by the SQL command" {
-			$parameter = [SqlCommandBuilder]::new($connection).GetFindCommand([Character], $record.Id).Item2[0]
-			$parameter.Name | Should -BeExactly "@ID"
-			$parameter.Value | Should -Be 1000
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetFindCommand([Character], $record.Id)
+			$parameters[0].Name | Should -BeExactly "@ID"
+			$parameters[0].Value | Should -Be 1000
 		}
 
 		It "should allow selecting a specific set of columns" {
-			$command = [SqlCommandBuilder]::new($connection).GetFindCommand([Character], $record.Id, "firstName").Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetFindCommand([Character], $record.Id, "firstName")
 			$command.Text | Should -BeLikeExactly 'SELECT "firstName"*'
 			$command.Text | Should -Not -BeLike "*gender*"
 			$command.Text | Should -Not -BeLike "*lastName*"
@@ -71,7 +71,7 @@ Describe "SqlCommandBuilder" {
 
 	Context "GetFindAllCommand" {
 		It "should return the SQL command to find all entities" {
-			$command = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character]).Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character])
 			$command.Text | Should -BeLikeExactly 'SELECT "*'
 			$command.Text | Should -Not -BeLike '*`**'
 			$command.Text | Should -BeLikeExactly '*FROM "main"."Characters"*'
@@ -79,13 +79,13 @@ Describe "SqlCommandBuilder" {
 		}
 
 		It "should also return an empty parameter collection" {
-			$parameters = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character]).Item2
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character])
 			$parameters | Should -BeNullOrEmpty
 		}
 
 		It "should allow sorting the results by a specific set of columns" {
 			$orderHints = [ordered]@{ gender = "Ascending"; fullName = "Descending" }
-			$command = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character], $orderHints).Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character], $orderHints)
 			$command.Text | Should -BeLikeExactly 'SELECT "*'
 			$command.Text | Should -Not -BeLike '*`**'
 			$command.Text | Should -BeLikeExactly '*FROM "main"."Characters"*'
@@ -93,7 +93,7 @@ Describe "SqlCommandBuilder" {
 		}
 
 		It "should allow selecting a specific set of columns" {
-			$command = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character], $null, "firstName").Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetFindAllCommand([Character], $null, "firstName")
 			$command.Text | Should -BeLikeExactly 'SELECT "firstName"*'
 			$command.Text | Should -Not -BeLike "*gender*"
 			$command.Text | Should -Not -BeLike "*lastName*"
@@ -103,13 +103,13 @@ Describe "SqlCommandBuilder" {
 
 	Context "GetInsertCommand" {
 		It "should return the SQL command to insert an entity" {
-			$command = [SqlCommandBuilder]::new($connection).GetInsertCommand($record).Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetInsertCommand($record)
 			$command.Text | Should -BeLikeExactly 'INSERT INTO "main"."Characters" (*'
 			$command.Text | Should -BeLikeExactly "*VALUES (*"
 		}
 
 		It "should also return the parameters used by the SQL command" {
-			$parameters = [SqlCommandBuilder]::new($connection).GetInsertCommand($record).Item2
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetInsertCommand($record)
 			$parameters | Should -HaveCount 3
 			$parameters["firstName"].Value | Should -BeExactly Cédric
 			$parameters["gender"].Value | Should -BeExactly DarkLord
@@ -119,14 +119,14 @@ Describe "SqlCommandBuilder" {
 
 	Context "GetUpdateCommand" {
 		It "should return the SQL command to update an entity" {
-			$command = [SqlCommandBuilder]::new($connection).GetUpdateCommand($record).Item1
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetUpdateCommand($record)
 			$command.Text | Should -BeLikeExactly 'UPDATE "main"."Characters"*'
 			$command.Text | Should -BeLikeExactly '*SET "*'
 			$command.Text | Should -BeLikeExactly '*WHERE "ID" = @ID'
 		}
 
 		It "should also return the parameters used by the SQL command" {
-			$parameters = [SqlCommandBuilder]::new($connection).GetUpdateCommand($record).Item2
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetUpdateCommand($record)
 			$parameters | Should -HaveCount 4
 			$parameters["ID"].Value | Should -Be 1000
 			$parameters["firstName"].Value | Should -BeExactly Cédric
@@ -135,7 +135,7 @@ Describe "SqlCommandBuilder" {
 		}
 
 		It "should allow updating a specific set of columns" {
-			$parameters = [SqlCommandBuilder]::new($connection).GetUpdateCommand($record, "firstName").Item2
+			$command, $parameters = [SqlCommandBuilder]::new($connection).GetUpdateCommand($record, "firstName")
 			$parameters | Should -HaveCount 2
 			$parameters["ID"].Value | Should -Be 1000
 			$parameters["firstName"].Value | Should -BeExactly Cédric